dms_backend/requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==3.2
django-cors-headers==3.14.0
django-filter==22.1
django-rest-durin==1.1.0
djangorestframework==3.12.0
djangorestframework-datatables==0.7.0
djangorestframework-link-header-pagination==0.1.1
filelock==3.8.2
humanize==4.6.0
pillow==10.3.0
platformdirs==2.6.0
psycopg2-binary==2.9.5
pypng==0.20220715.0
pytz==2022.7.1
qrcode==7.4.2
sqlparse==0.4.3
typing_extensions==4.12.2
virtualenv==20.17.1
